#263 moved filter, control and created base sz-graph components
