+++ src/lib/entity/detail/sz-entity-detail-graph/sz-standalone-graph.component.html
@@ -28,12 +28,12 @@
     (scaleChanged)="onGraphZoom($event)"
     [maxEntities]="maxEntities"></sz-relationship-network>
 
-    <sz-entity-detail-graph-control *ngIf="showMatchKeyControl" class="sz-entity-detail-graph-control"
+    <sz-graph-control *ngIf="showMatchKeyControl" class="sz-graph-control"
     [showLinkLabels]="_showMatchKeys"
     (optionChanged)="onOptionChange($event)"
-    ></sz-entity-detail-graph-control>
+    ></sz-graph-control>
 
-    <sz-entity-detail-graph-filter *ngIf="showFiltersControl" class="sz-entity-detail-graph-filter"
+    <sz-graph-filter *ngIf="showFiltersControl" class="sz-graph-filter"
       [ngStyle]="{'width.px': filterWidth}"
       [class.top-left]="filterControlPosition == 'top-left'"
       [class.top-right]="filterControlPosition == 'top-right'"
@@ -43,7 +43,7 @@
       (optionChanged)="onOptionChange($event)"
       [showDataSources]="filterShowDataSources"
       [showMatchKeys]="filterShowMatchKeys"
-    ></sz-entity-detail-graph-filter>
+    ></sz-graph-filter>
 
     <svg class="popout-icon"
       [class.top-left]="popOutIconPosition == 'top-left'"
