+++ src/lib/graph/sz-graph-control.component.scss
