+++ src/lib/graph/sz-graph-filter.component.scss
@@ -1,5 +1,5 @@
-@import "../../../scss/_variables";
-@import "../../../scss/_mixins";
+@import "../scss/_variables";
+@import "../scss/_mixins";
 
 :host {
   background-color: var(--sz-graph-filter-control-background-color);
