+++ src/lib/graph/sz-graph.component.html
@@ -1,0 +1,66 @@
+
+<div #graphContainer class="sz-graph-container">
+    <div *ngIf="showZoomControl" class="zoom-control-container"
+    [class.top-left]="zoomControlPosition == 'top-left'"
+    [class.top-right]="zoomControlPosition == 'top-right'"
+    [class.bottom-right]="zoomControlPosition == 'bottom-right'"
+    [class.bottom-left]="zoomControlPosition == 'bottom-left'">
+      <span><button class="zoom-out" (click)="zoomOut()"></button></span>
+      <input disabled="disabled" type="range" min="1" max="100" [value]="graphZoom" class="slider">
+      <span><button class="zoom-in" (click)="zoomIn()"></button></span>
+    </div>
+    <sz-relationship-network #graph class="sz-relationship-network-graph"
+      [class.filters-showing]="this.showFiltersControl"
+      svgViewBox="150 50 400 300"
+      svgPreserveAspectRatio="xMinYMid meet"
+      [entityIds]="graphIds"
+      [maxDegrees]="maxDegrees"
+      [buildOut]="buildOut"
+      [linkGravity]="5"
+      [highlight]="entityNodeColors"
+      [filter]="entityNodeFilters"
+      [includes]="entityMatchFilter"
+      [showLinkLabels]="_showMatchKeys"
+      (contextMenuClick)="onRightClick($event)"
+      (entityClick)="onEntityClick($event)"
+      (noResults)="onNoResults($event)"
+      (onDataLoaded)="onGraphDataLoaded($event)"
+      (scaleChanged)="onGraphZoom($event)"
+      [maxEntities]="maxEntities"></sz-relationship-network>
+  
+      <sz-graph-control *ngIf="showMatchKeyControl" class="sz-graph-control"
+      [showLinkLabels]="_showMatchKeys"
+      (optionChanged)="onOptionChange($event)"
+      ></sz-graph-control>
+  
+      <sz-graph-filter *ngIf="showFiltersControl" class="sz-graph-filter"
+        [ngStyle]="{'width.px': filterWidth}"
+        [class.top-left]="filterControlPosition == 'top-left'"
+        [class.top-right]="filterControlPosition == 'top-right'"
+        [class.bottom-right]="filterControlPosition == 'bottom-right'"
+        [class.bottom-left]="filterControlPosition == 'bottom-left'"
+        [showLinkLabels]="_showMatchKeys"
+        (optionChanged)="onOptionChange($event)"
+        [showDataSources]="filterShowDataSources"
+        [showMatchKeys]="filterShowMatchKeys"
+      ></sz-graph-filter>
+  
+      <svg class="popout-icon"
+        [class.top-left]="popOutIconPosition == 'top-left'"
+        [class.top-right]="popOutIconPosition == 'top-right'"
+        [class.bottom-right]="popOutIconPosition == 'bottom-right'"
+        [class.bottom-left]="popOutIconPosition == 'bottom-left'" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px"
+        viewBox="0 0 64 64"
+        (click)="onPopOutClick()"
+        *ngIf="showPopOutIcon"
+        enable-background="new 0 0 64 64" xml:space="preserve">
+        <g>
+          <g><g><polygon points="53,56 8,56 8,11 30,11 30,3 0,3 0,64 61,64 61,34 53,34"/></g></g>
+          <g><g><polygon points="42,0 50,8 33,25 39,31 56,14 64,23 64,0"/></g></g>
+        </g>
+      </svg>
+  </div>
+  <!--
+    filterControlPosition: "{{ filterControlPosition }}"<br/>
+    entityIds: {{ graphIds }} -->
+
